requirements.txt
# postgresql
psycopg2-binary==2.8.6

# ipython
ipython==7.19.0
ipython-genutils==0.2.0

# django restframework
djangorestframework==3.12.2

# swagger
drf-yasg==1.20.0

# allow different origin
django-cors-headers==3.6.0

# envrionment variable
django-environ==0.4.5